+++ src/app/app.module.ts
@@ -25,9 +25,9 @@ import { ResetemailComponent } from './components/resetemail/resetemail.componen
 import { Auth0LoginComponent } from './components/auth0login/auth0login.component';
 import { FileuploaderComponent } from './components/fileuploader/fileuploader.component';
 import { FileViewerComponent } from './components/file-viewer/file-viewer.component';
-import { ApplicantComponent } from './main/applicant/applicant.component';
-import { RecruiterComponent } from './main/recruiter/recruiter.component';
-import { MainComponent } from './main/main.component';
+import { ApplicantComponent } from './components/main/applicant/applicant.component';
+import { RecruiterComponent } from './components/main/recruiter/recruiter.component';
+import { MainComponent } from './components/main/main.component';
 
 @NgModule({
   declarations: [
